Functionally finished. Will make some minor changes
-Move device_detect functionality into keylogger.c
-Redirected log file to /log/keylogger.c
-Added include guards
-Updated Makefile

-TODO move keylogger.c to main
-TODO add descriptive comments
-TODO remove printf statements
