+++ Makefile
@@ -1,4 +1,4 @@
-OBJS = http.o ip.o
+OBJS = http.o ip.o key_util.o config.o
 
 main: main.c $(OBJS)
 	gcc main.c ip.o http.o -lpcap -o main
@@ -9,3 +9,8 @@ http.o: http.h http.c
 ip.o: ip.h ip.c
 	gcc -Wall -c ip.c
 
+key_util.o: key_util.h key_util.c
+gcc -Wall -c key_util.c
+
+config.o: config.h config.c
+gcc -Wall -c config.c
