+++ config.h
@@ -1,10 +1,13 @@
+#ifndef _CONFIG_H_
 #define _CONFIG_H_
+
 /*Config structure*/
-typedef struct Config {
+typedef struct Config{
    char *log_file;
    char *device_file;
-} Config;
+}Config;
 
 /*Free up space used by config struct*/
 void config_cleanup(Config *config);
 
+#endif
