--- device_detect.h
@@ -1,3 +1,0 @@
-#include "config.h"
-
-static char *get_keyboard_device_file_name();
